# Optimize runtime of test command