--- a/lib/learn_test.rb
+++ b/lib/learn_test.rb
@@ -1,5 +1,4 @@
 require 'fileutils'
-require 'faraday'
 require 'oj'
 require 'colorize'