--- a/lib/learn_test/runner.rb
+++ b/lib/learn_test/runner.rb
@@ -3,7 +3,6 @@ require 'yaml'
 module LearnTest
   class Runner
     attr_reader :repo, :options
-    SERVICE_URL = 'http://ironbroker-v2.flatironschool.com'
 
     def initialize(repo, options = {})
       @repo = repo
@@ -15,13 +14,15 @@ module LearnTest
       strategy.check_dependencies
       strategy.configure
       strategy.run
-      pid = Process.fork do
+      Process.detach(Process.fork do
+        require_relative 'reporter'
+
         if !help_option_present? && strategy.push_results? && !local_test_run?
-          push_results(strategy)
+          LearnTest::Reporter.report(strategy)
         end
+
         strategy.cleanup unless keep_results?
-      end
-      Process.detach(pid)
+      end)
     end
 
     def files
@@ -38,25 +39,6 @@ module LearnTest
 
     private
 
-    def augment_results!(results)
-      if File.exist?("#{FileUtils.pwd}/.learn")
-        dot_learn = YAML.load(File.read("#{FileUtils.pwd}/.learn"))
-
-        if !dot_learn['github'].nil?
-          results[:github] = dot_learn['github']
-        end
-      end
-
-      results[:ruby_platform] = RUBY_PLATFORM
-      results[:ide_container] = (ENV['IDE_CONTAINER'] == 'true')
-    end
-
-    def connection
-      @connection ||= Faraday.new(url: SERVICE_URL) do |faraday|
-        faraday.adapter  Faraday.default_adapter
-      end
-    end
-
     def strategies
       [
         LearnTest::Strategies::CSharpNunit,
@@ -69,21 +51,6 @@ module LearnTest
         LearnTest::Strategies::Mocha,
         LearnTest::Strategies::PythonUnittest
       ]
-    end
-
-    def push_results(strategy)
-      results = strategy.results
-      augment_results!(results)
-
-      begin
-        connection.post do |req|
-          req.url(strategy.service_endpoint)
-          req.headers['Content-Type'] = 'application/json'
-          req.body = Oj.dump(results, mode: :compat)
-        end
-      rescue Faraday::ConnectionFailed
-        puts 'There was a problem connecting to Learn. Not pushing test results.'.red
-      end
     end
 
     def help_option_present?