@joshrowley 2017-08-17
gonna probably test/QA some more tomorrow and then cut the new release on Monday